internal/templates/pgvector-rag/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.63.0
psycopg2-binary==2.9.9
pgvector==0.3.6
